requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
babel==2.16.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
filelock==3.16.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fsspec==2024.9.0
gradio_client==1.3.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
livereload==2.7.0
Mako==1.3.5
MarkupSafe==2.1.5
openai==1.46.0
packaging==24.1
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyTelegramBotAPI==4.23.0
python-telegram-bot==21.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
thread==2.0.5
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websockets==12.0
Werkzeug==3.0.3
